requirements.txt
pinecone
groq
langchain
langchain-community
sentence-transformers
python-dotenv
streamlit
tqdm